use ::deken::installpath rather than ::curent_plugin_loadpath
the $::current_plugin_loadpath is the path where the deken-plugin.tcl
is found.
this can be ~/pd-externals, but it can also be /usr/lib/pd/extra or
anything that is in the Pd-path.
it can also be ~/pd-externals/deken-plugin/, if this plugin ever becomes
complicated enough to ship in a sub-directory.

otoh, sys_staticpath holds a list of default search paths,
with the per-user path (~/pd-externals/ at the beginnint).

to make it easier to override this in the future, we put it into a
variable of our own