made "key_id" config work
rather than throw a nasty error.

also, fall back to "default-key" as found in the gpg-configuration.

FIXXME: currently we manually parse ~/.gnupg/gpg.conf, which is likely
to not work on non-linux (and is fragile on linux as well).
i haven't found a way to get the gpg-configuration from python-gnupg
(looking at the code, it seems there is no way)