added user-configuration: deken-plugin.conf
we allow for multiple configuration files, all of them are read:
config-files that are read earlier might have their values overwritten.

the user can put a configuration either directly besides the plugin-file,
or anywhere in their ${static-searchpath}/deken-plugin/.
(this might be handy if the plugin is installed system-wide, but the user
wants to add some user-specific configuration in their home).

the ${static-searchpath} is read in reversed order, so we first read the
system-wide installation paths, and finally the per-user installation paths.


configuration files are very simple. the first element of each line is the keyword,
the rest the value.
e.g.
<snip>
foo bar baz
</snip>
will store the list {bar baz} into ::deken::foo

LATER FIXXME: allowing any name gives the user the power to bork the system.
(at runtime)