upper-case the varargs variables
that's cosmetic but other variables are automatically uppercased in the
usage-message (e.g. '--version VERSION').

in order to avoid confusion between "deken package ..." and "deken upload package",
i think it is better to have all veriables uppercased.

e.g. "deken upload --version VERSION PACKAGE" makes it clear that "PACKAGE"
is not a literal...