Factored out apt-backend into separate plugin
it has the 'deken-xtra-...' prefix so it sorts alphabetically *after*
'deken-plugin'.
this allows to put both the 'deken-plugin.tcl' and the
'deken-xtra-apt-plugin.tcl' into the same directory and Pd will first load deken
and then the addon.

having the apt-backend available as a separate plugin has become necessary as
since the shipped-with-Pd deken has removed it from the core.

the backend itself is probably only useful for Pd installations via apt,
so it probably should be made available (only) via a Debian package.