only calculate distribution name when needed
calling lsb_release can take quite a while.
doing this on startup might considerably slow down the startup process.

since we only need to know the name for apt-found packages, we can defer
the calculation until the search-plugin is called.

in theory we could defer it even more (only calculate it the first time
the name is actually used - that is once we found our first package),
but the gain is little and code convolutes.